Fix message ordering problem for TCP clients.
TCP-based clients use one connection to acnetd to send commands and
receive command acks and ACNET traffic. There is a case where an reply
arrives before the ACK to the request is sent. This commit fixes this
problem by letting the data socket queue incoming ACNET packets until
an ACK is received from the command socket and sent to the client.

(Pulled from git.)